% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screener.R
\name{wkscreen_command}
\alias{wkscreen_command}
\title{Apply a command JSON and return the resulting response JSON.}
\usage{
wkscreen_command(screener, cmd_json)
}
\arguments{
\item{screener}{A screener handle from [wkscreen_new()].}

\item{cmd_json}{A command JSON string.}
}
\value{
The response as a JSON string.
}
\description{
Apply a command JSON and return the resulting response JSON.
}
